Move Plugin types into types.ts
