+++ packages/core/src/Repository.ts
@@ -9,6 +9,10 @@ import {
   type FileDataType,
   type MutativePatches,
   type NodeInfo,
+  type Plugin,
+  type PluginName,
+  type Plugins,
+  type PluginsData,
   isDirectoryId,
 } from "@/types";
 import type {
@@ -447,7 +451,7 @@ export interface RepositoryOptions<FT> {
   /** Provide a unique short id generator to create node ids. */
   createShortId?: CreateShortIdFunction;
 
-  plugins?: Partial<Record<PluginName, PluginOf["params"]>>;
+  plugins?: { [P in PluginName]?: Plugins[P]["params"] };
 }
 // #endregion
 // #region -- Drivers
@@ -493,32 +497,6 @@ export function registerPlugin<N extends PluginName>(
 ) {
   registeredPlugins[name] = plugin as never;
 }
-
-/** {@link Repository} plugin implementation function. */
-export interface Plugin<P = unknown> {
-  (this: Repository<any>, params: P | undefined): void;
-}
-/** `{params?,data?}` Declares integral types of a {@link Plugin}. */
-export interface PluginOf<P = undefined, D = unknown> {
-  /** Params passed when calling plugin. A `false` value disables the plugin. */
-  params?: P | boolean;
-  /** Type of the internal data stored in {@link Repository} by the plugin. */
-  data?: D;
-}
-/** `{"plug":{params?,data}}` Declare global {@link Plugin}s. */
-export interface Plugins {
-  // e.g. myPlugin: PluginOf<{foo?:"bar"|"baz"}>;
-  // "test": PluginOf<true, boolean>;
-}
-/** `{ [plugin]: plugin["data"] }` */
-export type PluginsData = {
-  /** Internal plugin data. One prop per registered plugin. */
-  [Prop in PluginName]?: Plugins[Prop]["data"];
-};
-
-/** Plugin name of a plugin registered in {@link Plugins} */
-export type PluginName = keyof Plugins & string;
-
 // #endregion
 
 export interface FsTransactions<FT> {
